Add role and role binding for sinkfilter controller.
